requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
Django==2.1.7
django-background-tasks==1.2.5
django-compat==1.0.15
djangorestframework==3.11.0
idna==2.9
psycopg2==2.8.5
pytz==2020.1
requests==2.23.0
six==1.14.0
urllib3==1.25.9
